Pass response details to `page` function (again)
After making use of the new version of Loom where you need to drop
down to `requestHandler` to get access to the response details, it
turns out it's kind of a pain. There are use cases where the normal
`page` behavior is what you want, but you still need to look at the
response details in order to decide what to do in the body (such as
switching on the status code in error pages).